+++ CHANGES.md
@@ -6,7 +6,8 @@ Change Log
 #### next release (8.0.0-alpha.77)
 
 - **Breaking changes**:
-  - `terria.error.raiseEvent` and `reaiseErrorToUser` have been replaced with `terria.raiseErrorToUser`
+  - `terria.error.raiseEvent` and `./raiseErrorToUser.ts` have been replaced with `terria.raiseErrorToUser`.
+  - `terria.error.raiseEvent` has been replaced with `terria.addErrorEventListener`
 
 * New Error handling using `Result` and `TerriaError` now applied to initial loading, `updateModelFromJson()`, `upsertModelFromJson()` and `Traits.fromJson()`. This means errors will propagate through these functions, and a stacktrace will be displayed.
 * New `terriaErrorNotification()` function, which wraps up error messages.